TOK-228 changing seq_no to u64

+++ libsovtoken/src/logic/parsers/parse_payment_response.rs
@@ -92,7 +92,7 @@ pub fn from_response(base : ParsePaymentResponse) -> Result<ParsePaymentReply, E
 
         let (address, amount) = unspent_output;
         let qualified_address: String = add_qualifer_to_address(&address);
-        let seq_no: i32 = base.result.tnx_meta_data.seq_no;
+        let seq_no: u64 = base.result.tnx_meta_data.seq_no;
         let txo = (TXO { address: qualified_address.to_string(), seq_no}).to_libindy_string()?;
         let utxo: UTXO = UTXO { payment_address: qualified_address, txo, amount, extra: "".to_string() };
